+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - the ability to remove fields entirely via the `modify` command (#142,!299)
 - a hint in the documentation how to filter for (non-)existence of fields (#141)
 
+### Fixed
+- the `Escape` key during an input prompt triggered after a `Confirm` prompt (e.g. a cancelled `quit`)
+
 
 ## [5.2.3] - 2024-12-15